Add Support for Header plugin and Fix Header Bug
The header content-type and datastar-request values
were being removed completely when adding new header
args to an action. This resulted in Laravel not seeing
the request body when calling $request->all() b/c the
content-type was default to 'text'.

I also added in the header plugin so I could just
set the header ONCE at the root of file instead of
cluttering up my action buttons with the same header
values again and again.